+++ pkg/combustion/registry_test.go
@@ -82,9 +82,9 @@ func TestWriteRegistryScript(t *testing.T) {
 
 	found := string(foundBytes)
 	assert.Contains(t, found, "cp $ARTEFACTS_DIR/registry/hauler /opt/hauler/hauler")
-	assert.Contains(t, found, "cp $ARTEFACTS_DIR/registry/embedded-registry.tar.zst /opt/hauler/")
+	assert.Contains(t, found, "cp $ARTEFACTS_DIR/registry/*-registry.tar.zst /opt/hauler/")
 	assert.Contains(t, found, "systemctl enable eib-embedded-registry.service")
-	assert.Contains(t, found, "ExecStartPre=/opt/hauler/hauler store load embedded-registry.tar.zst")
+	assert.Contains(t, found, "ExecStartPre=/bin/sh -c '/opt/hauler/hauler store load *-registry.tar.zst'")
 	assert.Contains(t, found, "ExecStart=/opt/hauler/hauler store serve registry -p 6545")
 }