+++ RELEASE_NOTES.md
@@ -11,6 +11,7 @@
 * [#319](https://github.com/suse-edge/edge-image-builder/issues/319) - Combustion fails when combustion directory content is larger than half of the RAM of the system
 * [#233](https://github.com/suse-edge/edge-image-builder/issues/233) - Use different Helm chart sources for development and production builds
 * [#95](https://github.com/suse-edge/edge-image-builder/issues/95) - Compressed images are not supported
+* [#343](https://github.com/suse-edge/edge-image-builder/issues/343) - Embedded Artifact Registry is memory bound
 
 ---